Enhance wake word detection, add fast TTS adapter, multi-voice support, and SFT scripts
- Expand wake word engine with configurable sensitivity, multi-keyword support, and audio feedback
- Add fast_tts adapter for lightweight voice agent inference
- Add multi_voice module for ClipCannon voice cloning pipeline
- Add CosyVoice SFT training scripts and jarvis service script
- Update conversation manager with improved turn handling and interruption logic
- Add local audio transport enhancements for real-time playback
- Update benchmarks app and documentation with latest results
- Add seed-TTS final benchmark results

+++ src/voiceagent/asr/streaming.py
@@ -24,7 +24,7 @@ class StreamingASR:
             self.model = faster_whisper.WhisperModel(
                 config.model_name,
                 device="cuda",
-                compute_type="int8",
+                compute_type="float32",
             )
         except Exception as e:
             raise ASRError(
